requirements.txt
torch
torchvision
opencv-python
numpy
scipy
scikit-image
gradio
Pillow
python-dotenv
requests
